add scheduler, worker, and executor services
