+++ Pipfile
@@ -14,6 +14,9 @@ ipython = "*"
 pymongo = {extras = ["srv"], version = "*"}
 sqlalchemy = "*"
 psycopg2-binary = "*"
+redis = "*"
+rq = "*"
+schedule = "*"
 
 [dev-packages]
