+++ Procfile
@@ -1,1 +1,3 @@
 web: uvicorn main:app --host 0.0.0.0 --port $PORT
+worker: python worker.py
+clock: python clock.py
